+++ mogong/src/main/java/hufs/team/mogong/image/service/dto/PreSignedUrlRequest.java
@@ -1,0 +1,16 @@
+package hufs.team.mogong.image.service.dto;
+
+import javax.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PUBLIC)
+public class PreSignedUrlRequest {
+
+	@NotBlank
+	private String extension;
+}
